+++ lib/graphql/resolvers/__mocks__/apolloServerContext.ts
@@ -1,12 +1,14 @@
-import TaxonomyAPI from '../../../../adapters/taxonomy'
-jest.mock('../../../../adapters/taxonomy')
+import TaxonomyAPI from '../../../adapters/taxonomy'
+jest.mock('../../../adapters/taxonomy')
 
 export const ctx = {
   headers: {
     token: 'token',
   },
   mydata: {
-    consents: {} as any,
+    consents: {
+      request: jest.fn(),
+    },
     getData: jest.fn(),
     saveData: jest.fn(),
   },
